Move delete organization to settings page
